pyproject.toml
[project]
name = "probe-evasion"
version = "0.1.0"
description = "Investigating whether language models can evade linear probes when informed about monitoring"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.45.0",
    "bitsandbytes>=0.41.0",
    "accelerate>=0.25.0",
    "scikit-learn>=1.3.0",
    "numpy",
    "pandas",
    "matplotlib",
    "seaborn",
    "pyyaml",
    "tqdm",
    "requests",
    "wikipedia-api",
    "streamlit>=1.37.0",
    "plotly>=5.18.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
